### 2026-01-08 #741 @danielvartan
logolink: An Interface for Running NetLogo Simulations from R
Submitting Author Name: Daniel Vartanian
Submitting Author Github Handle: @danielvartan
Repository:  https://github.com/danielvartan/logolink
Version submitted: 1.0.0
Submission type: Standard

Version accepted: TBD

---

```
Package: logolink
Title: An Interface for Running 'NetLogo' Simulations
Version: 1.0.0
        person(
            given = "Daniel",
            family = "Vartanian",
            role = c("aut", "cre", "ccp", "cph"),
            email = "danielvartan@proton.me",
            comment = c(ORCID = "0000-0001-7782-759X")
        )
    )
Description: An interface for 'NetLogo' <https://www.netlogo.org> that enables
    programmatic setup and execution of simulations. Designed to facilitate
    integrating 'NetLogo' models into reproducible workflows by creating and
    running 'BehaviorSpace' experiments and retrieving their results.
    https://github.com/danielvartan/logolink/
    checkmate (>= 2.3.3),
    cli (>= 3.6.5),
    fs (>= 1.6.6),
    glue (>= 1.8.0),
    janitor (>= 2.2.1),
    magrittr (>= 2.0.4),
    purrr (>= 1.2.0),
    readr (>= 2.0.0),
    tidyr (>= 1.3.1),
    stringr (>= 1.6.0),
    xml2 (>= 1.4.0)
    bslib (>= 0.9.0),
    colorspace (>= 2.1.2),
    covr (>= 3.6.5),
    curl (>= 7.0.0),
    ggplot2 (>= 4.0.1),
    ggimage (>= 0.3.4),
    ggtext (>= 0.1.2),
    here (>= 1.0.2),
    httr2 (>= 1.2.1),
    knitr (>= 1.50),
    lifecycle (>= 1.0.4),
    magick (>= 2.9.0),
    quarto (>= 1.5.1),
    ragg (>= 1.5.0),
    remotes (>= 2.5.0),
    rsvg (>= 2.7.0),
    scales (>= 1.4.0),
    spelling (>= 2.3.2),
    tibble (>= 3.3.0),
VignetteBuilder: quarto
```

`logolink` is an R package that simplifies setting up and running [NetLogo](https://www.netlogo.org/) simulations directly from R ([*scientific software wrappers*](https://devguide.ropensci.org/softwarereview_policies.html#package-categories)). It also includes routines for data tidying and parsing NetLogo data types ([*data munging*](https://devguide.ropensci.org/softwarereview_policies.html#package-categories)).

`logolink` was designed for researchers in complexity science, including ecology, epidemiology, social sciences, and economics, who work with agent-based modeling (ABM) and complex systems simulation.  It enables reproducible computational experiments and analysis across any domain where [NetLogo](https://www.netlogo.org/) models are employed.

While other R packages connect R to NetLogo, `logolink` is currently the only one that fully supports the latest NetLogo release ([*Broader in functionality*](https://devguide.ropensci.org/softwarereview_policies.html#overlap)). It is actively maintained, follows [tidyverse conventions](https://design.tidyverse.org/), and is designed to be simple and straightforward to use ([*More open in licensing or development practices*](https://devguide.ropensci.org/softwarereview_policies.html#overlap)).

For context, [`RNetLogo`](https://CRAN.R-project.org/package=RNetLogo) supports only older versions (up to 6.0.0, released in December 2016) and has not been updated since June 2017. [`nlrx`](https://CRAN.R-project.org/package=nlrx) offers a powerful framework for managing experiments and results, but [supports](https://docs.ropensci.org/nlrx/reference/supported_netlogo_versions.html) only up to NetLogo 6.3.0 (released in September 2022), requires additional system dependencies, uses its own internal conventions that diverge from NetLogo standards, and has [many unresolved issues](https://github.com/ropensci/nlrx/issues) ([*Actively maintained while alternatives are poorly or no longer actively maintained*](https://devguide.ropensci.org/softwarereview_policies.html#overlap)) ([*Better in usability and performance*](https://devguide.ropensci.org/softwarereview_policies.html#overlap)).

`logolink` complements these packages by prioritizing simplicity, offering finer control over output, ensuring full compatibility with NetLogo 7, and integrating seamlessly with modern R workflows.

```
── logolink 1.0.0 ───────────────────────────────────────────────────────────────

✔ Package is already on CRAN.
✔ has a 'codemeta.json' file.
✔ has a 'contributing' file.
✔ uses 'roxygen2'.
✔ 'DESCRIPTION' has a URL field.
✔ 'DESCRIPTION' has a BugReports field.
✔ Package has at least one HTML vignette
✔ All functions have examples.
✔ Package has continuous integration checks.
✔ Package coverage is 100%.
✔ All goodpractice linters passed.
✔ R CMD check found no errors.
✔ R CMD check found no warnings.
ℹ Function names are duplicated in other packages
ℹ Examples should not use `\dontrun` unless really necessary.

ℹ Current status:
✔ A truly lovely set of check results!

[...]

✖ The following 3 function names are duplicated in other packages:
  • - `create_experiment` from previsionio, SigOptR, cometr, SqueakR
  • - `read_experiment` from Rtrack
  • - `run_experiment` from CAISEr, R6causal, banditsCI, calmr
```

There is minimal overlap in function names. These function names follow the [Tidyverse Style Guide](https://style.tidyverse.org/functions.html#naming), making the package usage more intuitive and user-friendly.

Some examples must use `\dontrun` because CRAN runners do not have NetLogo installations or allow internet connectivity.
