+++ tests/testthat.R
@@ -26,6 +26,7 @@ test_check("logolink")
 # devtools:test()
 # goodpractice::gp()
 # lintr::lint_package()
+# pkgcheck::pkgcheck()
 # pkgdown::build_article("x")
 # pkgdown::build_favicons(overwrite = TRUE)
 # pkgdown::build_reference()
