### 2026-01-08 57a1e46
Add CONTRIBUTING guide
