+++ .github/CONTRIBUTING.md
@@ -1,0 +1,97 @@
+# Contributing to `logolink`
+
+<!-- This CONTRIBUTING.md was adapted from https://gist.github.com/peterdesmet/e90a1b0dc17af6c12daf6e8b2f044e7c -->
+
+First of all, thank you for considering a contribution to `logolink`! Your interest and involvement are what make it rewarding for us (the project maintainers) to continue developing and improving the package.
+
+`logolink` is an open source project, maintained by people who care. We are not directly funded to do so.
+
+[repo]: https://github.com/danielvartan/logolink
+[issues]: https://github.com/danielvartan/logolink/issues
+[new_issue]: https://github.com/danielvartan/logolink/issues/new
+[website]: https://danielvartan.github.io/logolink
+[citation]: https://danielvartan.github.io/logolink/authors.html
+
+## How You Can Contribute
+
+There are several ways you can contribute to this project. If you want to know more about why and how to contribute to open source projects like this one, see this [Open Source Guide](https://opensource.guide/how-to-contribute/).
+
+### Give It a Star ⭐
+
+[![GitHub repo stars](https://img.shields.io/github/stars/danielvartan/logolink)](https://github.com/danielvartan/logolink/)
+
+If you find `logolink` helpful, please consider starring the repository on GitHub. It helps others discover the project and shows your support!
+
+### Share `logolink` with Others 📣
+
+Think `logolink` is useful? Let others discover it, by telling them in person, via social media.
+
+### Sponsor `logolink` 💖
+
+[![GitHub Sponsor](https://img.shields.io/static/v1?label=Sponsor&message=%E2%9D%A4&logo=GitHub&color=%23fe8e86)](https://github.com/sponsors/danielvartan)
+
+If you want to support `logolink` and its development, consider [sponsoring the project](https://github.com/sponsors/danielvartan). Your support helps us maintain and improve the project, and it allows us to dedicate more time to its development. Don't forget to mention `logolink` in your sponsorship message.
+
+### Cite `logolink` in Your Work 📝
+
+Are you using `logolink` in a publication or project? Please consider [citing it][citation]. Citations help support the project and demonstrate its impact to the community.
+
+### Propose an Idea 💡
+
+Have an idea for a new `logolink` feature? Take a look at the [documentation][website] and [issue list][issues] to see if it isn't included or suggested yet. If not, suggest your idea as an [issue on GitHub][issues]. While we can't promise to implement your idea, it helps to:
+
+- Explain in detail how it would work.
+- Keep the scope as narrow as possible.
+
+See below if you want to contribute code for your idea as well.
+
+### Report a Bug 🐛
+
+Using `logolink` and discovered a bug? That's annoying! Don't let others have the same experience and report it as an [issue on GitHub][new_issue] so we can fix it. A good bug report makes it easier for us to do so, so please include:
+
+* The content of `utils::sessionInfo()`.
+* Any details about your local setup that might be helpful in troubleshooting.
+* Detailed steps to reproduce the bug (Tip: Use the [reprex](https://reprex.tidyverse.org/) package).
+
+### Improve the Documentation 📖
+
+Noticed a typo on the website? Think a function could use a better example? Good documentation makes all the difference, so your help to improve it is very welcome!
+
+#### The Website
+
+The [website][website] is generated with [`pkgdown`](http://pkgdown.r-lib.org/). That means we don't have to write any HTML: Content is pulled together from documentation in the code, vignettes, [Markdown](https://guides.github.com/features/mastering-markdown/) files, the package `DESCRIPTION` and `_pkgdown.yml` settings. If you know your way around `pkgdown`, you can [propose a file change](https://help.github.com/articles/editing-files-in-another-user-s-repository/) to improve documentation.
+
+#### Function Documentation
+
[…]
