+++ pkgdown.yml
@@ -4,7 +4,7 @@ pkgdown_sha: ~
 articles:
   logolink: logolink.html
   visualizing-the-netlogo-world: visualizing-the-netlogo-world.html
-last_built: 2026-01-08T09:32Z
+last_built: 2026-01-08T19:37Z
 urls:
   reference: https://danielvartan.github.io/logolink/reference
   article: https://danielvartan.github.io/logolink/articles
