+++ R/.post-render.R
@@ -44,7 +44,7 @@ file |>
 
 # Update Package Versions in `DESCRIPTION` -----
 
-# update_pkg_versions()
+update_pkg_versions()
 
 # Update Package Year in `LICENSE` and `inst/CITATION` -----
 
