+++ NEWS.md
@@ -1,3 +1,5 @@
+# logolink 1.0.1.9000 (development version)
+
 # logolink 1.0.0
 
 ### Breaking Changes
