+++ README.md
@@ -5,35 +5,11 @@
 <!-- Install the package before rendering this file: `devtools::install()` -->
 
 <!-- badges: start -->
-[![Project Status: Active - The project has reached a stable, usable
-state and is being actively
-developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
-[![CRAN Status
-Badge](https://www.r-pkg.org/badges/version-ago/logolink)](https://cran.r-project.org/package=logolink)
-[![DOI
-badge](https://img.shields.io/badge/doi-10.32614/CRAN.package.logolink-1284C5.svg)](https://doi.org/10.32614/CRAN.package.logolink)
-[![](http://cranlogs.r-pkg.org/badges/grand-total/logolink)](https://cran.r-project.org/package=logolink)
-[![R build
-status](https://github.com/danielvartan/logolink/workflows/R-CMD-check/badge.svg)](https://github.com/danielvartan/logolink/actions)
-[![CRAN
-checks](https://badges.cranchecks.info/summary/logolink.svg)](https://cran.r-project.org/web/checks/check_results_logolink.html)
-[![FAIR checklist
-badge](https://img.shields.io/badge/fairsoftwarechecklist.net--00a7d9)](https://fairsoftwarechecklist.net/v0.2/?f=31&a=32113&i=32322&r=133)
-[![](https://codecov.io/gh/danielvartan/logolink/branch/main/graph/badge.svg)](https://app.codecov.io/gh/danielvartan/logolink)
-[![GNU GPLv3
-License](https://img.shields.io/badge/license-GPLv3-bd0000.svg)](https://www.gnu.org/licenses/gpl-3.0)
-[![FAIR checklist
-badge](https://img.shields.io/badge/fairsoftwarechecklist.net--00a7d9)](https://fairsoftwarechecklist.net/v0.2/?f=31&a=32113&i=32322&r=133)
-[![fair-software.eu](https://img.shields.io/badge/fair--software.eu-%E2%97%8F%20%20%E2%97%8F%20%20%E2%97%8F%20%20%E2%97%8F%20%20%E2%97%8F-green)](https://fair-software.eu)
-[![Contributor Covenant 3.0 Code of
-Conduct](https://img.shields.io/badge/Contributor%20Covenant-3.0-4baaaa.svg)](https://www.contributor-covenant.org/version/3/0/code_of_conduct/)
-<!-- badges: end -->
-
-## Overview
+logolink 1.0.0.9000 (development version) \## Overview
 
 `logolink` is an [R](https://www.r-project.org/) package that simplifies
 setting up and running [NetLogo](https://www.netlogo.org/) simulations
-directly from R. It provides a modern, intuitive interface that follows
+from R. It provides a modern, intuitive interface that follows
 [tidyverse
 principles](https://tidyverse.tidyverse.org/articles/manifesto.html) and
 integrates seamlessly with the [tidyverse
@@ -133,7 +109,7 @@ Volterra ([1926](https://www.nature.com/articles/118558a0)). Since this
 model comes bundled with NetLogo, no download is required.
 
 We’ll use
-[`find_netlogo_home`](https://danielvartan.github.io/logolink/reference/find_netlogo_home.html)
+[`find_netlogo_home()`](https://danielvartan.github.io/logolink/reference/find_netlogo_home.html)
 function to locate the NetLogo installation directory, then build the
 path to the model file:
 
@@ -154,7 +130,7 @@ To run the model from R, we’ll need to setup an experiment. We can do
 this by setting a
 [BehaviorSpace](https://docs.netlogo.org/behaviorspace.html) experiment
 with the
[…]

[…]
