+++ news/index.md
@@ -1,5 +1,7 @@
 # Changelog
 
+## logolink 1.0.1.9000 (development version)
+
 ## logolink 1.0.0
 
 CRAN release: 2026-01-08
