+++ reference/run_experiment.md
@@ -318,18 +318,18 @@ Other BehaviorSpace functions:
 #> ✔ Running model [20.1s]
 #> 
 #> ℹ Gathering metadata
-#> ✔ Gathering metadata [26ms]
+#> ✔ Gathering metadata [18ms]
 #> 
 #> ℹ Processing table output
-#> ✔ Processing table output [21ms]
+#> ✔ Processing table output [15ms]
 #> 
 #> ℹ The experiment run produced the following messages:
 #> 
-#> Jan 08, 2026 10:29:37 PM java.util.prefs.FileSystemPreferences$1 run
+#> Jan 08, 2026 11:22:01 PM java.util.prefs.FileSystemPreferences$1 run
 #> INFO: Created user preferences directory.
 #> $metadata
 #> $metadata$timestamp
-#> [1] "2026-01-08 22:29:38 GMT"
+#> [1] "2026-01-08 23:22:02 GMT"
 #> 
 #> $metadata$netlogo_version
 #> [1] "7.0.3"
@@ -377,7 +377,7 @@ Other BehaviorSpace functions:
       experiment = "Wolf Sheep Simple model analysis"
     )
 #> ℹ Running model
-#> ✔ Running model [17.7s]
+#> ✔ Running model [19s]
 #> 
 #> ℹ Gathering metadata
 #> ✔ Gathering metadata [14ms]
@@ -387,7 +387,7 @@ Other BehaviorSpace functions:
 #> 
 #> $metadata
 #> $metadata$timestamp
-#> [1] "2026-01-08 22:29:59 GMT"
+#> [1] "2026-01-08 23:22:23 GMT"
 #> 
 #> $metadata$netlogo_version
 #> [1] "7.0.3"
