+++ index.md
@@ -1,6 +1,6 @@
 # logolink
 
-logolink 1.0.0.9000 (development version) \## Overview
+## Overview
 
 `logolink` is an [R](https://www.r-project.org/) package that simplifies
 setting up and running [NetLogo](https://www.netlogo.org/) simulations
@@ -208,7 +208,7 @@ library(dplyr)
 results |> glimpse()
 #> List of 2
 #>  $ metadata:List of 6
-#>   ..$ timestamp       : POSIXct[1:1], format: "2026-01-08 18:42:07"
+#>   ..$ timestamp       : POSIXct[1:1], format: "2026-01-08 20:18:42"
 #>   ..$ netlogo_version : chr "7.0.3"
 #>   ..$ output_version  : chr "2.0"
 #>   ..$ model_file      : chr "Wolf Sheep Simple 5.nlogox"
@@ -225,7 +225,7 @@ results |> glimpse()
 #>   ..$ energy_gain_from_sheep: num [1:110110] 5 5 5 5 5 5 5 5 5 5 ...
 #>   ..$ step                  : num [1:110110] 0 1 2 3 4 5 6 7 8 9 ...
 #>   ..$ count_wolves          : num [1:110110] 5 5 5 5 5 5 5 5 5 5 ...
-#>   ..$ count_sheep           : num [1:110110] 500 500 498 497 495 494 494 492 490 487 ...
+#>   ..$ count_sheep           : num [1:110110] 500 500 499 499 496 495 495 493 492 490 ...
 ```
 
 If you already have a file with experiment results, you can read it into
