### 2026-01-11 265bf5c
Add `SystemRequirements` field to DESCRIPTION
