+++ NEWS.md
@@ -1,5 +1,7 @@
 # logolink 1.0.1.9000 (development version)
 
+- Patch changes were added to the documentation.
+
 # logolink 1.0.0
 
 ### Breaking Changes
