+++ DESCRIPTION
@@ -53,6 +53,7 @@ Suggests:
     spelling (>= 2.3.2),
     tibble (>= 3.3.0),
     testthat (>= 3.3.0)
+SystemRequirements: NetLogo >= 7.0.1
 VignetteBuilder: quarto
 Config/testthat/edition: 3
 Encoding: UTF-8
