+++ reference/run_experiment.md
@@ -315,21 +315,21 @@ Other BehaviorSpace functions:
       setup_file = setup_file
     )
 #> ℹ Running model
-#> ✔ Running model [20.1s]
+#> ✔ Running model [20s]
 #> 
 #> ℹ Gathering metadata
-#> ✔ Gathering metadata [18ms]
+#> ✔ Gathering metadata [19ms]
 #> 
 #> ℹ Processing table output
 #> ✔ Processing table output [15ms]
 #> 
 #> ℹ The experiment run produced the following messages:
 #> 
-#> Jan 08, 2026 11:22:01 PM java.util.prefs.FileSystemPreferences$1 run
+#> Jan 11, 2026 12:50:09 AM java.util.prefs.FileSystemPreferences$1 run
 #> INFO: Created user preferences directory.
 #> $metadata
 #> $metadata$timestamp
-#> [1] "2026-01-08 23:22:02 GMT"
+#> [1] "2026-01-11 00:50:10 GMT"
 #> 
 #> $metadata$netlogo_version
 #> [1] "7.0.3"
@@ -377,7 +377,7 @@ Other BehaviorSpace functions:
       experiment = "Wolf Sheep Simple model analysis"
     )
 #> ℹ Running model
-#> ✔ Running model [19s]
+#> ✔ Running model [18s]
 #> 
 #> ℹ Gathering metadata
 #> ✔ Gathering metadata [14ms]
@@ -387,7 +387,7 @@ Other BehaviorSpace functions:
 #> 
 #> $metadata
 #> $metadata$timestamp
-#> [1] "2026-01-08 23:22:23 GMT"
+#> [1] "2026-01-11 00:50:31 GMT"
 #> 
 #> $metadata$netlogo_version
 #> [1] "7.0.3"
