+++ news/index.md
@@ -2,6 +2,8 @@
 
 ## logolink 1.0.1.9000 (development version)
 
+- Patch changes were added to the documentation.
+
 ## logolink 1.0.0
 
 CRAN release: 2026-01-08
