+++ .github/workflows/check-standard.yaml
@@ -17,10 +17,10 @@ jobs:
           - {os: ubuntu-latest, r: 'devel', http-user-agent: 'release'}
           - {os: ubuntu-latest, r: 'release'}
           - {os: ubuntu-latest, r: 'oldrel-1'}
-    name: ${{matrix.config.os}} (${{matrix.config.r}})
-    runs-on: ${{matrix.config.os}}
+    name: ${{ matrix.config.os }} (${{ matrix.config.r }})
+    runs-on: ${{ matrix.config.os }}
     env:
-      GH_TOKEN: ${{secrets.GITHUB_TOKEN}}
+      GH_TOKEN: ${{ secrets.GITHUB_TOKEN }}
       R_KEEP_PKG_SOURCE: yes
     steps:
       - name: Check out code
@@ -35,8 +35,8 @@ jobs:
       - name: Set up R
         uses: r-lib/actions/setup-r@v2
         with:
-          r-version: ${{matrix.config.r}}
-          http-user-agent: ${{matrix.config.http-user-agent}}
+          r-version: ${{ matrix.config.r }}
+          http-user-agent: ${{ matrix.config.http-user-agent }}
           use-public-rspm: true
 
       - name: Install R dependencies
@@ -48,8 +48,8 @@ jobs:
       - name: Set up TinyTeX
         uses: r-lib/actions/setup-tinytex@v2
 
-      # We skip --run-dontrun and --run-donttest due to
-      # LaTeX issues on Windows.
+      # --run-dontrun and --run-donttest
+      # are skipped on Windows due to LaTeX issues.
       - name: Run R CMD check
         if: runner.os == 'Windows'
         uses: r-lib/actions/check-r-package@v2
