### 2026-01-13 bd5c032
Add `output_dir`  as argument to `run_experiment()`
