+++ tests/testthat/test-run_experiment.R
@@ -14,11 +14,7 @@ testthat::test_that("`run_experiment()` | General test", {
   run_experiment(
     model_path = model_file_1,
     setup_file = setup_file_1,
-    experiment = NULL,
-    output = "table",
-    other_arguments = NULL,
-    timeout = Inf,
-    tidy_output = TRUE
+    output = "table"
   ) |>
     checkmate::expect_list() |>
     suppressMessages() |>
@@ -103,7 +99,8 @@ testthat::test_that("`run_experiment()` | Error test", {
     output = "table",
     other_arguments = NULL,
     timeout = Inf,
-    tidy_output = TRUE
+    tidy_output = TRUE,
+    output_dir = temp_dir()
   ) |>
     testthat::expect_error()
 
@@ -116,7 +113,8 @@ testthat::test_that("`run_experiment()` | Error test", {
     output = "table",
     other_arguments = NULL,
     timeout = Inf,
-    tidy_output = TRUE
+    tidy_output = TRUE,
+    output_dir = temp_dir()
   ) |>
     testthat::expect_error()
 
@@ -129,7 +127,8 @@ testthat::test_that("`run_experiment()` | Error test", {
     output = "table",
     other_arguments = NULL,
     timeout = Inf,
-    tidy_output = TRUE
+    tidy_output = TRUE,
+    output_dir = temp_dir()
   ) |>
     testthat::expect_error()
 
@@ -142,7 +141,8 @@ testthat::test_that("`run_experiment()` | Error test", {
     output = "table",
     other_arguments = NULL,
     timeout = Inf,
-    tidy_output = TRUE
+    tidy_output = TRUE,
+    output_dir = temp_dir()
   ) |>
     testthat::expect_error()
 
@@ -155,7 +155,8 @@ testthat::test_that("`run_experiment()` | Error test", {
     output = "table",
     other_arguments = NULL,
     timeout = Inf,
-    tidy_output = TRUE
+    tidy_output = TRUE,
+    output_dir = temp_dir()
   ) |>
     testthat::expect_error()
 
@@ -168,7 +169,8 @@ testthat::test_that("`run_experiment()` | Error test", {
     output = "table",
     other_arguments = NULL,
     timeout = Inf,
-    tidy_output = TRUE
+    tidy_output = TRUE,
+    output_dir = temp_dir()
   ) |>
     testthat::expect_error()
 
@@ -181,7 +183,8 @@ testthat::test_that("`run_experiment()` | Error test", {
     output = 1,
     other_arguments = NULL,
     timeout = Inf,
-    tidy_output = TRUE
+    tidy_output = TRUE,
+    output_dir = temp_dir()
   ) |>
     testthat::expect_error()
 
@@ -192,7 +195,8 @@ testthat::test_that("`run_experiment()` | Error test", {
     output = character(),
     other_arguments = NULL,
     timeout = Inf,
-    tidy_output = TRUE
+    tidy_output = TRUE,
+    output_dir = temp_dir()
   ) |>
     testthat::expect_error()
 
@@ -205,7 +209,8 @@ testthat::test_that("`run_experiment()` | Error test", {
     output = "test",
     other_arguments = NULL,
     timeout = Inf,
-    tidy_output = TRUE
+    tidy_output = TRUE,
+    output_dir = temp_dir()
   ) |>
     testthat::expect_error()
 
@@ -218,7 +223,8 @@ testthat::test_that("`run_experiment()` | Error test", {
     output = "table",
     other_arguments = 1,
     timeout = Inf,
-    tidy_output = TRUE
+    tidy_output = TRUE,
+    output_dir = temp_dir()
   ) |>
     testthat::expect_error()
 
[…]
