+++ tests/testthat/test-run_experiment-sod.R
@@ -49,11 +49,7 @@ testthat::test_that("`run_experiment()` | Population Density (Runtime)", {
     model_path |>
     run_experiment(
       setup_file = setup_file,
-      experiment = NULL,
-      output = c("table", "spreadsheet", "lists", "statistics"),
-      other_arguments = NULL,
-      timeout = Inf,
-      tidy_output = TRUE
+      output = c("table", "spreadsheet", "lists", "statistics")
     ) |>
     suppressMessages() |>
     suppressWarnings()
