+++ news/index.md
@@ -3,6 +3,9 @@
 ## logolink 1.0.1.9000 (development version)
 
 - Patch changes were added to the documentation.
+- [`run_experiment()`](https://danielvartan.github.io/logolink/reference/run_experiment.md)
+  now have an `output_dir` parameter to specify the directory where
+  NetLogo experiment output files are stored.
 
 ## logolink 1.0.0
 
