+++ CITATION.cff
@@ -188,7 +188,7 @@ references:
     email: lionel@posit.co
   year: '2026'
   doi: 10.32614/CRAN.package.purrr
-  version: '>= 1.2.0'
+  version: '>= 1.2.1'
 - type: software
   title: readr
   abstract: 'readr: Read Rectangular Text Data'
@@ -225,7 +225,7 @@ references:
     given-names: Maximilian
   year: '2026'
   doi: 10.32614/CRAN.package.tidyr
-  version: '>= 1.3.1'
+  version: '>= 1.3.2'
 - type: software
   title: stringr
   abstract: 'stringr: Simple, Consistent Wrappers for Common String Operations'
@@ -255,7 +255,7 @@ references:
     email: jeroenooms@gmail.com
   year: '2026'
   doi: 10.32614/CRAN.package.xml2
-  version: '>= 1.5.0'
+  version: '>= 1.5.1'
 - type: software
   title: bslib
   abstract: 'bslib: Custom ''Bootstrap'' ''Sass'' Themes for ''shiny'' and ''rmarkdown'''
@@ -399,7 +399,7 @@ references:
     orcid: https://orcid.org/0000-0002-6485-8781
   year: '2026'
   doi: 10.32614/CRAN.package.ggimage
-  version: '>= 0.3.4'
+  version: '>= 0.3.5'
 - type: software
   title: ggtext
   abstract: 'ggtext: Improved Text Rendering Support for ''ggplot2'''
@@ -444,7 +444,7 @@ references:
     email: hadley@posit.co
   year: '2026'
   doi: 10.32614/CRAN.package.httr2
-  version: '>= 1.2.1'
+  version: '>= 1.2.2'
 - type: software
   title: knitr
   abstract: 'knitr: A General-Purpose Package for Dynamic Report Generation in R'
@@ -458,7 +458,7 @@ references:
     orcid: https://orcid.org/0000-0003-0645-5666
   year: '2026'
   doi: 10.32614/CRAN.package.knitr
-  version: '>= 1.50'
+  version: '>= 1.51'
 - type: software
   title: lifecycle
   abstract: 'lifecycle: Manage the Life Cycle of your Package Functions'
@@ -475,7 +475,7 @@ references:
     orcid: https://orcid.org/0000-0003-4757-117X
   year: '2026'
   doi: 10.32614/CRAN.package.lifecycle
-  version: '>= 1.0.4'
+  version: '>= 1.0.5'
 - type: software
   title: magick
   abstract: 'magick: Advanced Graphics and Image-Processing in R'
@@ -627,5 +627,5 @@ references:
     email: hadley@posit.co
   year: '2026'
   doi: 10.32614/CRAN.package.testthat
-  version: '>= 3.3.0'
+  version: '>= 3.3.1'
 
