+++ DESCRIPTION
@@ -27,23 +27,23 @@ Imports:
     glue (>= 1.8.0),
     janitor (>= 2.2.1),
     magrittr (>= 2.0.4),
-    purrr (>= 1.2.0),
+    purrr (>= 1.2.1),
     readr (>= 2.1.6),
-    tidyr (>= 1.3.1),
+    tidyr (>= 1.3.2),
     stringr (>= 1.6.0),
-    xml2 (>= 1.5.0)
+    xml2 (>= 1.5.1)
 Suggests:
     bslib (>= 0.9.0),
     colorspace (>= 2.1.2),
     covr (>= 3.6.5),
     curl (>= 7.0.0),
     ggplot2 (>= 4.0.1),
-    ggimage (>= 0.3.4),
+    ggimage (>= 0.3.5),
     ggtext (>= 0.1.2),
     here (>= 1.0.2),
-    httr2 (>= 1.2.1),
-    knitr (>= 1.50),
-    lifecycle (>= 1.0.4),
+    httr2 (>= 1.2.2),
+    knitr (>= 1.51),
+    lifecycle (>= 1.0.5),
     magick (>= 2.9.0),
     quarto (>= 1.5.1),
     ragg (>= 1.5.0),
@@ -52,7 +52,7 @@ Suggests:
     scales (>= 1.4.0),
     spelling (>= 2.3.2),
     tibble (>= 3.3.0),
-    testthat (>= 3.3.0)
+    testthat (>= 3.3.1)
 SystemRequirements: NetLogo >= 7.0.1
 VignetteBuilder: quarto
 Config/testthat/edition: 3
