### 2026-01-14 9daf632
Add pkgcheck workflow
