+++ tests/testthat/test-run_experiment-vsa.R
@@ -9,12 +9,6 @@ testthat::test_that("`run_experiment()` | Sensitivity Experiment", {
   testthat::skip_if(is.na(find_netlogo_console()))
   testthat::skip_if(Sys.getenv("LOGOLINK_TEST_NETLOGO") == "FALSE")
 
-  testthat::skip_if(
-    find_netlogo_console() |>
-      suppressMessages() |>
-      identical("")
-  )
-
   model_path <-
     find_netlogo_home() |>
     file.path(
