### 2026-01-14 4e6f7f4
Remove `pkgcheck.yaml`
