+++ README.md
@@ -23,8 +23,6 @@ checks](https://badges.cranchecks.info/summary/logolink.svg)](https://cran.r-pro
 [![](https://codecov.io/gh/danielvartan/logolink/branch/main/graph/badge.svg)](https://app.codecov.io/gh/danielvartan/logolink)
 [![FAIR checklist
 badge](https://img.shields.io/badge/fairsoftwarechecklist.net--00a7d9)](https://fairsoftwarechecklist.net/v0.2/?f=31&a=32113&i=32322&r=133)
-[![FAIR checklist
-badge](https://img.shields.io/badge/fairsoftwarechecklist.net--00a7d9)](https://fairsoftwarechecklist.net/v0.2/?f=31&a=32113&i=32322&r=133)
 [![fair-software.eu](https://img.shields.io/badge/fair--software.eu-%E2%97%8F%20%20%E2%97%8F%20%20%E2%97%8F%20%20%E2%97%8F%20%20%E2%97%8F-green)](https://fair-software.eu)
 [![GNU GPLv3
 License](https://img.shields.io/badge/license-GPLv3-bd0000.svg)](https://www.gnu.org/licenses/gpl-3.0)
@@ -168,7 +166,6 @@ metrics to collect, and the number of runs to perform.
 setup_file <- create_experiment(
   name = "Wolf Sheep Simple Model Analysis",
   repetitions = 10,
-  sequential_run_order = TRUE,
   run_metrics_every_step = TRUE,
   setup = "setup",
   go = "go",
@@ -177,7 +174,6 @@ setup_file <- create_experiment(
     'count wolves',
     'count sheep'
   ),
-  run_metrics_condition = NULL,
   constants = list(
     "number-of-sheep" = 500,
     "number-of-wolves" = list(
@@ -240,7 +236,7 @@ library(dplyr)
 results |> glimpse()
 #> List of 2
 #>  $ metadata:List of 6
-#>   ..$ timestamp       : POSIXct[1:1], format: "2026-01-14 01:24:37"
+#>   ..$ timestamp       : POSIXct[1:1], format: "2026-01-18 19:14:38"
 #>   ..$ netlogo_version : chr "7.0.3"
 #>   ..$ output_version  : chr "2.0"
 #>   ..$ model_file      : chr "Wolf Sheep Simple 5.nlogox"
@@ -257,7 +253,7 @@ results |> glimpse()
 #>   ..$ energy_gain_from_sheep: num [1:110110] 5 5 5 5 5 5 5 5 5 5 ...
 #>   ..$ step                  : num [1:110110] 0 1 2 3 4 5 6 7 8 9 ...
 #>   ..$ count_wolves          : num [1:110110] 5 5 5 5 5 5 5 5 5 5 ...
-#>   ..$ count_sheep           : num [1:110110] 500 497 496 496 495 491 490 489 486 486 ...
+#>   ..$ count_sheep           : num [1:110110] 500 499 498 496 494 494 492 490 489 487 ...
 ```
 
 If you already have a file with experiment results, you can read it into
