+++ vignettes/logolink.qmd
@@ -79,7 +79,6 @@ Now that we have our environment set up, let's start by creating an experiment. 
 setup_file <- create_experiment(
   name = "Population Density (Runtime)",
   repetitions = 10,
-  sequential_run_order = TRUE,
   run_metrics_every_step = TRUE,
   time_limit = 1000,
   setup = 'setup',
