+++ vignettes/visualizing-the-netlogo-world.qmd
@@ -102,9 +102,6 @@ Here's where things get interesting. We want to capture data of every sheep, wol
 ```{r}
 setup_file <- create_experiment(
   name = "Wolf Sheep Simple Model Analysis",
-  repetitions = 1,
-  sequential_run_order = TRUE,
-  run_metrics_every_step = FALSE,
   setup = "setup",
   go = "go",
   time_limit = 500,
