### 2026-01-18 a1605ce
Ignore `readr` and `xml2` when updating packaging versions
