+++ DESCRIPTION
@@ -28,10 +28,10 @@ Imports:
     janitor (>= 2.2.1),
     magrittr (>= 2.0.4),
     purrr (>= 1.2.1),
-    readr (>= 2.1.6),
+    readr (>= 2.0.0),
     tidyr (>= 1.3.2),
     stringr (>= 1.6.0),
-    xml2 (>= 1.5.1)
+    xml2 (>= 1.5.0)
 Suggests:
     bslib (>= 0.9.0),
     colorspace (>= 2.1.2),
