### 2026-01-18 3d02eec
Remove patch versions from dependencies
