### 2026-01-18 9755538
Add `rlang (>= 1.1.7)` to Suggests
