+++ DESCRIPTION
@@ -20,39 +20,40 @@ URL: https://danielvartan.github.io/logolink/,
 BugReports: https://github.com/danielvartan/logolink/issues/
 Depends: R (>= 4.4)
 Imports:
-    checkmate (>= 2.3),
-    cli (>= 3.6),
-    dplyr (>= 1.1),
-    fs (>= 1.6),
-    glue (>= 1.8),
-    janitor (>= 2.2),
-    magrittr (>= 2.0),
-    purrr (>= 1.2),
-    readr (>= 2.0),
-    tidyr (>= 1.3),
-    stringr (>= 1.6),
-    xml2 (>= 1.5)
+    checkmate (>= 2.3.3),
+    cli (>= 3.6.5),
+    dplyr (>= 1.1.4),
+    fs (>= 1.6.6),
+    glue (>= 1.8.0),
+    janitor (>= 2.2.1),
+    magrittr (>= 2.0.4),
+    purrr (>= 1.2.1),
+    readr (>= 2.1.6),
+    tidyr (>= 1.3.2),
+    stringr (>= 1.6.0),
+    xml2 (>= 1.5.1)
 Suggests:
-    bslib (>= 0.9),
-    colorspace (>= 2.1),
-    covr (>= 3.6),
-    curl (>= 7.0),
-    ggplot2 (>= 4.0),
-    ggimage (>= 0.3),
-    ggtext (>= 0.1),
-    here (>= 1.0),
-    httr2 (>= 1.2),
-    knitr (>= 1.5),
-    lifecycle (>= 1.0),
-    magick (>= 2.9),
-    quarto (>= 1.5),
-    ragg (>= 1.5),
-    remotes (>= 2.5),
-    rsvg (>= 2.7),
-    scales (>= 1.4),
-    spelling (>= 2.3),
-    tibble (>= 3.3),
-    testthat (>= 3.3)
+    bslib (>= 0.9.0),
+    colorspace (>= 2.1.2),
+    covr (>= 3.6.5),
+    curl (>= 7.0.0),
+    ggplot2 (>= 4.0.1),
+    ggimage (>= 0.3.5),
+    ggtext (>= 0.1.2),
+    here (>= 1.0.2),
+    httr2 (>= 1.2.2),
+    knitr (>= 1.51),
+    lifecycle (>= 1.0.5),
+    magick (>= 2.9.0),
+    quarto (>= 1.5.1),
+    ragg (>= 1.5.0),
+    remotes (>= 2.5.0),
+    rlang (>= 1.1.7),
+    rsvg (>= 2.7.0),
+    scales (>= 1.4.0),
+    spelling (>= 2.3.2),
+    tibble (>= 3.3.0),
+    testthat (>= 3.3.1)
 SystemRequirements: NetLogo >= 7.0.1
 VignetteBuilder: quarto
 Config/testthat/edition: 3
