+++ reference/run_experiment.md
@@ -343,17 +343,17 @@ Other BehaviorSpace functions:
       setup_file = setup_file
     )
 #> ℹ Running model
-#> ✔ Running model [21.3s]
+#> ✔ Running model [19.9s]
 #> 
 #> ℹ Gathering metadata
-#> ✔ Gathering metadata [20ms]
+#> ✔ Gathering metadata [14ms]
 #> 
 #> ℹ Processing table output
-#> ✔ Processing table output [16ms]
+#> ✔ Processing table output [15ms]
 #> 
 #> $metadata
 #> $metadata$timestamp
-#> [1] "2026-01-18 23:34:51 GMT"
+#> [1] "2026-01-18 23:55:53 GMT"
 #> 
 #> $metadata$netlogo_version
 #> [1] "7.0.3"
@@ -401,17 +401,17 @@ Other BehaviorSpace functions:
       experiment = "Wolf Sheep Simple model analysis"
     )
 #> ℹ Running model
-#> ✔ Running model [18s]
+#> ✔ Running model [17.9s]
 #> 
 #> ℹ Gathering metadata
-#> ✔ Gathering metadata [14ms]
+#> ✔ Gathering metadata [15ms]
 #> 
 #> ℹ Processing table output
-#> ✔ Processing table output [9ms]
+#> ✔ Processing table output [8ms]
 #> 
 #> $metadata
 #> $metadata$timestamp
-#> [1] "2026-01-18 23:35:13 GMT"
+#> [1] "2026-01-18 23:56:13 GMT"
 #> 
 #> $metadata$netlogo_version
 #> [1] "7.0.3"
