### 2026-01-19 9a61e69
Update `pkgcheck.yaml`
