+++ R/.post-render.R
@@ -44,13 +44,7 @@ file |>
 
 # Update Package Versions in `DESCRIPTION` -----
 
-# update_pkg_versions(
-#   ignore = c(
-#     "dplyr",
-#     "readr",
-#     "xml2"
-#   )
-# )
+update_pkg_versions()
 
 # Update Package Year in `LICENSE` and `inst/CITATION` -----
 
