### 2026-01-19 dc91fe7
Remove `pkgcheck.yaml`
