+++ README.md
@@ -16,8 +16,6 @@ Downloads](https://cranlogs.r-pkg.org/badges/grand-total/logolink)](https://dani
 badge](https://img.shields.io/badge/doi-10.32614/CRAN.package.logolink-1284C5.svg)](https://doi.org/10.32614/CRAN.package.logolink)
 [![R build
 status](https://github.com/danielvartan/logolink/workflows/R-CMD-check.yaml/badge.svg)](https://github.com/danielvartan/logolink/actions)
-[![R build
-status](https://github.com/danielvartan/logolink/workflows/pkgcheck.yaml/badge.svg)](https://github.com/danielvartan/logolink/actions)
 [![CRAN
 checks](https://badges.cranchecks.info/summary/logolink.svg)](https://cran.r-project.org/web/checks/check_results_logolink.html)
 [![](https://codecov.io/gh/danielvartan/logolink/branch/main/graph/badge.svg)](https://app.codecov.io/gh/danielvartan/logolink)
@@ -236,7 +234,7 @@ library(dplyr)
 results |> glimpse()
 #> List of 2
 #>  $ metadata:List of 6
-#>   ..$ timestamp       : POSIXct[1:1], format: "2026-01-18 19:14:38"
+#>   ..$ timestamp       : POSIXct[1:1], format: "2026-01-18 21:52:23"
 #>   ..$ netlogo_version : chr "7.0.3"
 #>   ..$ output_version  : chr "2.0"
 #>   ..$ model_file      : chr "Wolf Sheep Simple 5.nlogox"
@@ -253,7 +251,7 @@ results |> glimpse()
 #>   ..$ energy_gain_from_sheep: num [1:110110] 5 5 5 5 5 5 5 5 5 5 ...
 #>   ..$ step                  : num [1:110110] 0 1 2 3 4 5 6 7 8 9 ...
 #>   ..$ count_wolves          : num [1:110110] 5 5 5 5 5 5 5 5 5 5 ...
-#>   ..$ count_sheep           : num [1:110110] 500 499 498 496 494 494 492 490 489 487 ...
+#>   ..$ count_sheep           : num [1:110110] 500 497 494 490 489 487 486 483 481 479 ...
 ```
 
 If you already have a file with experiment results, you can read it into
