+++ reference/run_experiment.md
@@ -346,14 +346,14 @@ Other BehaviorSpace functions:
 #> ✔ Running model [19.9s]
 #> 
 #> ℹ Gathering metadata
-#> ✔ Gathering metadata [14ms]
+#> ✔ Gathering metadata [21ms]
 #> 
 #> ℹ Processing table output
-#> ✔ Processing table output [15ms]
+#> ✔ Processing table output [16ms]
 #> 
 #> $metadata
 #> $metadata$timestamp
-#> [1] "2026-01-18 23:55:53 GMT"
+#> [1] "2026-01-19 00:56:51 GMT"
 #> 
 #> $metadata$netlogo_version
 #> [1] "7.0.3"
@@ -401,17 +401,17 @@ Other BehaviorSpace functions:
       experiment = "Wolf Sheep Simple model analysis"
     )
 #> ℹ Running model
-#> ✔ Running model [17.9s]
+#> ✔ Running model [18s]
 #> 
 #> ℹ Gathering metadata
-#> ✔ Gathering metadata [15ms]
+#> ✔ Gathering metadata [14ms]
 #> 
 #> ℹ Processing table output
 #> ✔ Processing table output [8ms]
 #> 
 #> $metadata
 #> $metadata$timestamp
-#> [1] "2026-01-18 23:56:13 GMT"
+#> [1] "2026-01-19 00:57:12 GMT"
 #> 
 #> $metadata$netlogo_version
 #> [1] "7.0.3"
