### 2026-01-19 8d4762d
Fix "Visualizing the NetLogo World" vignette
