### 2026-01-19 618e66a
Add `pkgcheck.yaml`
