+++ .github/workflows/pkgcheck.yaml
@@ -1,0 +1,22 @@
+on:
+  push:
+    branches: [main, master]
+  workflow_dispatch:
+
+name: pkgcheck.yaml
+permissions: read-all
+
+concurrency:
+  group: ${{ github.workflow }}-${{ github.head_ref }}
+  cancel-in-progress: true
+
+jobs:
+  pkgcheck:
+    runs-on: ubuntu-latest
+    permissions:
+      issues: write
+    steps:
+      - name: Check package with rOpenSci's pkgcheck system
+        uses: ropensci-review-tools/pkgcheck-action@main
+        with:
+          post-to-issue: false
