### 2026-01-19 cc2c592
Add `workflow_run` condition to `pkgcheck.yaml`
