### 2026-01-19 a3abda4
Fix `pkgcheck.yaml`
