### 2026-01-21 43f9d99
Update pkgcheck workflow
