+++ .github/workflows/pkgcheck.yaml
@@ -1,7 +1,7 @@
 on:
   workflow_run:
-    workflows: "R-CMD-check.yaml"
-    types: completed
+    workflows: ["R-CMD-check.yaml"]
+    types: [completed]
     branches: [main, master]
 
 name: pkgcheck.yaml
@@ -17,15 +17,6 @@ jobs:
     permissions:
       issues: write
     steps:
-      - name: Check if all workflows succeeded
-        run: |
-          if [ "${{ github.event.workflow_run.conclusion }}" != "success" ]; then
-            echo "Previous workflow failed"
-
-            exit 1
-          fi
-        shell: bash
-
       - name: Check package with rOpenSci's pkgcheck system
         uses: ropensci-review-tools/pkgcheck-action@main
         with:
