### 2026-01-22 #741 @danielvartan
Hi @jhollist,

There is considerable overlap between [`nlrx`](https://docs.ropensci.org/nlrx/) and [`logolink`](https://danielvartan.github.io/logolink/), but the packages have different focuses, so I believe they could complement each other.

`nlrx` concentrates on experimental design and analysis, while `logolink` takes a closer approach to the NetLogo API. In that sense, `logolink` could handle the heavy lifting for `nlrx`, as you suggested. However, I'm not certain this would work in practice, since `nlrx` uses its own approach for setting up experiments. What I can say is that `logolink`'s connection with the NetLogo API is now feature-complete, so this integration is at least possible.

That said, this integration is not yet available to users, as it would require some adaptations to the data structures used in `nlrx`. As of today, the two packages don't compete because `logolink` works only with NetLogo 7 and above, while `nlrx` works only with older versions of NetLogo. If `nlrx` offered support for NetLogo 7, it could be used to achieve similar results to `logolink`, making them competing packages for the same tasks, although with `nlrx` offering less control over the output.

Another area of potential complementarity is parsing tools and API connections with other projects. However, the package doesn't currently offer many features in this area, so I'm unsure whether these tools would be useful for `nlrx`.

I do plan to add analysis tools in the future, but that's not a priority at the moment. My current focus is on adding an interface with the Java API to [control NetLogo's interface directly](https://docs.netlogo.org/controlling), similar to what [`pynetlogo`](https://pynetlogo.readthedocs.io/en/latest/index.html) offers.
