+++ DESCRIPTION
@@ -20,9 +20,9 @@ URL: https://danielvartan.github.io/logolink/,
 BugReports: https://github.com/danielvartan/logolink/issues/
 Depends: R (>= 4.4)
 Imports:
-    checkmate (>= 2.3.3),
+    checkmate (>= 2.3.4),
     cli (>= 3.6.5),
-    dplyr (>= 1.1.4),
+    dplyr (>= 1.2.0),
     fs (>= 1.6.6),
     glue (>= 1.8.0),
     janitor (>= 2.2.1),
@@ -31,13 +31,13 @@ Imports:
     readr (>= 2.1.6),
     tidyr (>= 1.3.2),
     stringr (>= 1.6.0),
-    xml2 (>= 1.5.1)
+    xml2 (>= 1.5.2)
 Suggests:
-    bslib (>= 0.9.0),
+    bslib (>= 0.10.0),
     colorspace (>= 2.1.2),
     covr (>= 3.6.5),
     curl (>= 7.0.0),
-    ggplot2 (>= 4.0.1),
+    ggplot2 (>= 4.0.2),
     ggimage (>= 0.3.5),
     ggtext (>= 0.1.2),
     here (>= 1.0.2),
@@ -51,8 +51,8 @@ Suggests:
     rsvg (>= 2.7.0),
     scales (>= 1.4.0),
     spelling (>= 2.3.2),
-    tibble (>= 3.3.0),
-    testthat (>= 3.3.1)
+    tibble (>= 3.3.1),
+    testthat (>= 3.3.2)
 SystemRequirements: NetLogo >= 7.0.1
 VignetteBuilder: quarto
 Config/testthat/edition: 3
