### 2026-02-11 c0ead44
Add brand.yml to Suggests
