### 2026-03-25 #741 @danielvartan
Thank you @jmaspons and @bitbacchus for your thorough reviews and kind words. I'm really glad you both liked the package!

Your suggestions are great and I'll start implementing them right away. I'll write a detailed response here for each one.

I'm a little short on time these days, but I believe I'll be able to respond by April 20th.
