### 2026-03-25 #741 @jhollist
All,

Great progress on this.  Really happy to see this moving forward.

@danielvartan no problem with your proposed timeline.  I'll keep my eye open for your revisions.

@jmaspons and @bitbacchus I have one more request for you.  Keep an eye our for the revisions.  Once those are in can you indicate your final approval with this form: https://devguide.ropensci.org/approval2template.html.
