### 2026-04-28 #741 @danielvartan
Hi @jhollist,

Sorry for the delay. I'm going through a job transition, which has been cutting into my time for other projects.

Brazil has an extended holiday weekend later this week, and I'm planning to use it to catch up here. I think I can clear all the pending items in a solid day of work.

Thanks for your patience.
