+++ reference/find_netlogo_console.md
@@ -40,6 +40,6 @@ Other system functions:
 ``` r
 # \dontrun{
   find_netlogo_console()
-#> /opt/NetLogo 7.0.3/bin/NetLogo
+#> /opt/NetLogo 7.0.4/bin/NetLogo
 # }
 ```
