+++ reference/get_netlogo_shape.md
@@ -86,11 +86,6 @@ the file paths to the downloaded NetLogo shapes as
 #> 1.06K
 
   shape |> image_read_svg() |> image_ggplot()
-#> Warning: `aes_string()` was deprecated in ggplot2 3.0.0.
-#> ℹ Please use tidy evaluation idioms with `aes()`.
-#> ℹ See also `vignette("ggplot2-in-packages")` for more information.
-#> ℹ The deprecated feature was likely used in the magick package.
-#>   Please report the issue at <https://github.com/ropensci/magick/issues>.
 
 # }
 
