+++ reference/find_netlogo_home.md
@@ -39,6 +39,6 @@ Other system functions:
 ``` r
 # \dontrun{
   find_netlogo_home()
-#> /opt/NetLogo 7.0.3
+#> /opt/NetLogo 7.0.4
 # }
 ```
