+++ dev/articles/visualizing-the-netlogo-world.md
@@ -1,0 +1,383 @@
+# Visualizing the NetLogo World
+
+## Overview
+
+This vignette demonstrates how to capture and visualize
+[NetLogo](https://www.netlogo.org/) simulations at specific time steps
+using `logolink` and [`ggplot2`](https://ggplot2.tidyverse.org/). You’ll
+learn how to extract agent positions, render them as publication-ready
+figures, and create animations showing simulation dynamics over time.
+
+We’ll work with Wilensky’s [Wolf Sheep
+Simple](https://www.netlogoweb.org/launch#https://www.netlogoweb.org/assets/modelslib/IABM%20Textbook/chapter%204/Wolf%20Sheep%20Simple%205.nlogox)
+model, a classic predator-prey simulation based on the [Lotka-Volterra
+equations](https://danielvartan.github.io/lotka-volterra/) formulated by
+Alfred J. Lotka
+([1925](http://archive.org/details/elementsofphysic017171mbp)) and Vito
+Volterra ([1926](https://www.nature.com/articles/118558a0)). This model
+ships with NetLogo, so no separate download is needed. By the end of
+this vignette, you’ll have both static plots and an animated
+[GIF](https://en.wikipedia.org/wiki/GIF) showing the simulation evolving
+over time.
+
+This guide assumes familiarity with [NetLogo](https://www.netlogo.org)
+(version 7.0.1 or above) and R programming, particularly the
+[tidyverse](https://tidyverse.org/) ecosystem.
+
+## Setting the Stage
+
+First, let’s load the packages we’ll need:
+
+``` r
+
+library(logolink)
+
+library(dplyr)
+library(ggplot2)
+library(ggimage)
+library(ggtext)
+library(here)
+library(magick)
+library(magrittr)
+library(ragg)
+library(stringr)
+library(tidyr)
+```
+
+If any of these are missing from your system, install them with:
+
+``` r
+
+install.packages(
+  c(
+    "dplyr",
+    "ggplot2",
+    "ggimage",
+    "ggtext",
+    "logolink",
+    "here",
+    "magick",
+    "magrittr",
+    "ragg",
+    "stringr",
+    "tidyr"
+  )
+)
+```
+
+Next, we need to locate the model. We’ll use the
+[`find_netlogo_home()`](https://danielvartan.github.io/logolink/reference/find_netlogo_home.html)
+function to find the NetLogo installation, then navigate to the model
+file:
+
+``` r
+
+model_path <-
+  find_netlogo_home() |>
+  file.path(
+    "models",
+    "IABM Textbook",
+    "chapter 4",
+    "Wolf Sheep Simple 5.nlogox"
+  )
+```
+
+We’ll also need the turtle shapes to make our plots look nice. We’ll use
+the
+[`get_netlogo_shape()`](https://danielvartan.github.io/logolink/reference/get_netlogo_shape.html)
+function to download turtle [SVG](https://en.wikipedia.org/wiki/SVG)
+image files from the
+[LogoShapes](https://github.com/danielvartan/logoshapes) project:
+
+``` r
+
+sheep_shape <- get_netlogo_shape("sheep")
+```
+
+``` r
+
+wolf_shape <- get_netlogo_shape("wolf")
+```
+
+## Running the Simulation
+
+Here’s where things get interesting. We want to capture data of every
+sheep, wolf, and patch at regular intervals. Let’s set up an experiment
+with
[…]
