+++ dev/reference/read_experiment.md
@@ -1,0 +1,161 @@
+# Read NetLogo BehaviorSpace Experiment output
+
+`read_experiment()` reads NetLogo
+[BehaviorSpace](https://docs.netlogo.org/behaviorspace.html) experiment
+output files as [tidy data
+frames](https://r4ds.hadley.nz/data-tidy.html). It automatically detects
+the output format
+([Table](https://docs.netlogo.org/behaviorspace.html#table-output),
+[Spreadsheet](https://docs.netlogo.org/behaviorspace.html#spreadsheet-output),
+[Lists](https://docs.netlogo.org/behaviorspace.html#lists-output), or
+[Stats](https://docs.netlogo.org/behaviorspace.html#statistics-output))
+and parses the data accordingly. The function also extracts metadata
+from the files.
+
+Only version 2.0 (NetLogo 6.4 and later) of BehaviorSpace output files
+is supported.
+
+## Usage
+
+``` r
+read_experiment(file, tidy_output = TRUE)
+```
+
+## Arguments
+
+- file:
+
+  A [`character`](https://rdrr.io/r/base/character.html) string
+  specifying the path to the
+  [BehaviorSpace](https://docs.netlogo.org/behaviorspace.html) output
+  [CSV](https://en.wikipedia.org/wiki/Comma-separated_values) file.
+
+- tidy_output:
+
+  (optional) A [`logical`](https://rdrr.io/r/base/logical.html) flag
+  indicating whether to tidy the output data frames. If `TRUE`, output
+  data frames are arranged according to [tidy data
+  principles](https://r4ds.hadley.nz/data-tidy.html). If `FALSE`, only
+  the default transformations from
+  [`read_delim()`](https://readr.tidyverse.org/reference/read_delim.html)
+  and
+  [`clean_names()`](https://sfirke.github.io/janitor/reference/clean_names.html)
+  are applied to the output data (default: `TRUE`).
+
+## Value
+
+A [`list`](https://rdrr.io/r/base/list.html) containing the experiment
+results. The list includes the following elements, depending on the
+output file provided:
+
+- `metadata`: A [`list`](https://rdrr.io/r/base/list.html) with metadata
+  about the experiment run (present in all cases).
+
+- `table`: A
+  [`tibble`](https://tibble.tidyverse.org/reference/tibble.html) with
+  the results of the
+  [`table`](https://docs.netlogo.org/behaviorspace.html#table-output)
+  output.
+
+- `spreadsheet`: A [`list`](https://rdrr.io/r/base/list.html) with the
+  results of the
+  [`spreadsheet`](https://docs.netlogo.org/behaviorspace.html#spreadsheet-output)
+  output containing two elements:
+
+  - `statistics`: A
+    [`tibble`](https://tibble.tidyverse.org/reference/tibble.html) with
+    data from the output first section.
+
+  - `data`: A
+    [`tibble`](https://tibble.tidyverse.org/reference/tibble.html) with
+    data from the output second section.
+
+- `lists`: A
+  [`tibble`](https://tibble.tidyverse.org/reference/tibble.html) with
+  the results of the
+  [`lists`](https://docs.netlogo.org/behaviorspace.html#lists-output)
+  output.
+
+- `statistics`: A
+  [`tibble`](https://tibble.tidyverse.org/reference/tibble.html) with
+  the results of the
+  [`statistics`](https://docs.netlogo.org/behaviorspace.html#statistics-output)
+  output.
+
+## See also
+
+Other BehaviorSpace functions:
[…]
