### 2026-06-07 danielvartan/logolink#16 @danielvartan
Review heuristics
> [!NOTE]
> `logolink` [rOpenSci](https://ropensci.org/) software-review: https://github.com/ropensci/software-review/issues/741#issuecomment-3993016097 (@jmaspons)

> Link with NetLogo
>
> The package did not find the NetLogo installation directory in my system, but it's well documented how to solve the problem.
>
> The find_netlogo_home() points to ~/.netlogo in my system. This is a folder created by NetLogo to save configurations and, therefore, quite common in environments running NetLogo, but not related to the installation folder. Just use a case-sensitive filter or, even better, check for the executables to confirm that the proposed folder indeed contains NetLogo.
Perhaps add a break [here](https://github.com/danielvartan/logolink/blob/c4f5cc948b36d94d1a55c16ab209f1919b392108/R/find_netlogo_home.R#L87) and sort find_netlogo_home.common_paths as needed, otherwise the result is overwritten unnecessarily.
>
> On Linux, I would add code to check for /usr/share/applications/NetLogo-7.0.3.desktop (created by install.sh bundled with the NetLogo download) and guess the NetLogo home from there too.
>
> The documentation explaining how to define the NetLogo home from in the section "Setting the NetLogo Installation Path" in run_experiment() would better fit in find_netlogo(), which is shorter in content and probably only used by users when there are problems finding the NetLogo home.
