### 2026-06-07 danielvartan/logolink#24 @danielvartan
Review `utils-mocks.R`
> [!NOTE]
> `logolink` [rOpenSci](https://ropensci.org/) software-review: https://github.com/ropensci/software-review/issues/741#issuecomment-4118226239 (@bitbacchus)

> **utils-mocks.R naming and purpose**
>
> I found utils-mocks.R confusing. The file references ?testthat::local_mocked_bindings, which suggests temporary function replacement for testing.
> However, the functions defined there appear to be permanent aliases rather than actual mocks.
>
> This creates a mismatch between naming/documentation and functionality.
> I would suggest either:
>
> - renaming the file (e.g. to utils-aliases.R), or
> - clarifying its purpose and removing the reference to local_mocked_bindings
