### 2026-06-07 danielvartan/logolink#26 @danielvartan
Clean up or document non-essential files
> [!NOTE]
> `logolink` [rOpenSci](https://ropensci.org/) software-review: https://github.com/ropensci/software-review/issues/741#issuecomment-4118226239 (@bitbacchus)

> Repository hygiene / contributor clarity
>
> - ` .vscode` directory
> The .vscode folder contains editor-specific configuration and is not needed for users or contributors.
> I would recommend removing it from the repository.
>
> Hidden helper scripts (`R/.setup.R`, `R/.post-render.R`)
> These scripts are useful for documentation/site generation but not obvious to contributors.
> Adding a short comment explaining their purpose and noting that they are excluded from the package build (via .Rbuildignore) would improve transparency.
