### 2026-06-07 danielvartan/logolink#27 @danielvartan
Add links to PRs, issues, or commits in `NEWS.md`
> [!NOTE]
> `logolink` [rOpenSci](https://ropensci.org/) software-review: https://github.com/ropensci/software-review/issues/741#issuecomment-3993016097 (@jmaspons)

> The repo has no Pull Requests. It’s a good practice to aggregate related commits and describe the issues that the changes want to address. It helps to understand code decisions and to review the development. Also, using PR is useful to catch errors before merging to main as the checks, code coverage and other actions get triggered in the branch.
>
> To help users to dig in to the changes of the new versions, I would add links to PR, issues or commits in the NEWS.
